itoa func and fix linker

+++ src/shell.c
@@ -18,7 +18,7 @@ void PrintHelp(){
 }
 
 int readline(char *buf, int size){
-  int idx = 0;
+  unsigned int idx = 0;
   char c = uart_getc();
   uart_send(c);
 
@@ -30,18 +30,23 @@ int readline(char *buf, int size){
       uart_send(c); // need to recv the echo back
   }
   buf[idx] = '\0';
-  return 0;
+  return idx;
 }
 
 
 void ShellLoop(){
   char buf[MAX_SIZE];
   while(1){
-    readline(buf, sizeof(buf));
+    unsigned int size = readline(buf, sizeof(buf));
+    if (size == 0){
+      uart_puts("# ");
+      continue;
+    } 
     if(strcmp("help", buf) == 0) PrintHelp();
     else if(strcmp("hello", buf) == 0) uart_puts("Hello World!\n");
     else if(strcmp("reboot", buf) == 0) uart_puts("Rebooting...\n");
     else uart_puts("Unknown command\n");
+    
 
     uart_puts("# ");
   }
